requirements.txt
web3==6.11.0
python-okx==0.2.3